Restructure project to use Poetry
The project now uses poetry as a dependency manager. As a result of this
it is now setup like a python package.

The `monopoly` class is also now again using Python's random as it
produces better random numbers than C's `rand` does. When testing with
over 1,000,000 turns using the C `rand` function the output probability
data was the same every time.