Allow for more turns to be tested with the C ext
Previously, the maximum number of turns per game in the C extension was
stored as an int, as were the results for each space. Now we use a long
long which allows for up to 2**63-1. This is per game, so the total
number of turns that you can test depends on how many cores your
machine has access to.