Rename app directory to monopoly
This changes the package folder name back to monopoly. This is done
because it is possible to use a __main__.py file as an entry point to
build binaries. This works with PyInstaller and will work with
PyOxidizer. I think it will work for Nuitka but will see. I think this
is slightly better and it results in not needing a wrapper file as an
entry point.