Always use multiprocessing when running simulation
Rather than only using multiprocessing for mutli-core mode, use it all
the time. While this does have a slight performance penalty for lower
turn runs, it allows the spinner to always work.

The way it was before it seemed the c extension code was blocking the
thread from allowing the spinner to get a chance to update when running
in single core mode. After attempts to get around this in other ways, I
think this is the simplest.