Add support for Nuitka builds
Unfortunately this does not add support for multi-core support with
Nuitka. Despite my efforts I was unable to get multiprocessing working
when running the application that was built from Nuitka. For now
multi-core is turned off for Nuitka but hopefully I can figure it out
and get it working in the future.

P.S. To force the Nuitka build to use multiprocessing, set the
environment variable FORCE_NUITKA_MULTI at runtime.