Add the ability to choose pure python simulation
This adds a command line flag that indicates the simulation should run
using the pure python `Monopoly` class.

The refactoring done in order to accomplish this also fixes the problem
where the fallback text would get printed to the screen multiple times
when run in parallel (this was a byproduct of how multiprocessing works
when using the 'spawn' start method). Now printing the text occurs
inside a function that is called only once.