Revert simplifying Nuitka build process
I thought Nuitka was building the `monopoly.so` file itself when
building the binary. However, it seems it is just copying the one that
is located in the top level 'app' package if it is there. When it isn't
there it just omits it and the final binary is built with no C extension
'Monopoly' class. Changing this back to how it originally worked
guarantees the C extension is always there.