Replace Makefile with install_monopoly.py script
The primary motivation of this commit is to have a cross platform script
file. Make, while it could be installed and run on Windows, didn't
actually work since the commands were all unix based. By having the
script be python, it should now work on both Mac and Windows (hopefully
Linux as well).