Update build-system in pyproject.toml
With these changes, the build system now uses setuptools exclusively to
handle installing monopoly-probabilities. This is faster and also
supports using the setup.py just for building, which is done in
scriptopoly for building the c extension for multiple commands.

The decision to not upgrade to the modern poetry build backend was made
because of the need for using setup.py for building. The poetry-core
build backend was also much faster than what was previously used, but it
does not allow for a way to *just* build the extension. To do that the
setup.py file and setuptools are still needed. Having one tool handle
both pip installs and project builds seems simpler and thus was chosen.

While the tool.poetry sections are still currently in the pyproject.toml
file, they may be removed in a future update. Currently, the only
function poetry is serving is as a dependency manager. However, it is
not needed by scriptopoly and thus is only a convenience.