Fix scriptopoly command
This adds a `pth` file into the virtual environments site-packages
directory. This tells python where to look to find the `script.py` file.

Doing it this way avoids needing to add the script file into the
package. This saves the trouble of having to make sure the packaging
tools all ignore the script file when building the binaries.