Add archive-binaries command to script
This will now handle the archiving of the binary builds to a zip or
tar.gz file itself.

Also, it was becoming necessary to add a way to specify a custom
location for the dist directory. So this takes care of making that
change to any of the prior build commands that use the dist directory as
well as the new archive command.