Add plotly for bar chart generation
This uses plotly to generate a bar chart with the results of the
simulation. Looks great, not that hard to do, but....WAY to big of a
dependency. And that doesn't even include the functionality to produce
a png, thats another large dependency thats required to do that with
plotly. I'm checking this in so it can be in the history but
unfortunately this is not a viable solution for this project.