Continue to work on Pygal integration
This gets the chart to a basically finished version. It adds colors to
the bars based on the actual monopoly board color. It also begins to
address most of the issues with freezing the binary with the chart
capabilities built-in. With this, both the PyInstaller and the Nuitka
binaries are working when I build them locally on my mac.

This also updates `setup.py` to match the project metadata from
`pyproject.toml`. I forgot to do this in the previous release.