List each file for artifact upload
Upload Artifact will include the entire directory tree when uploading
artifacts. Now that the archive is not located in the root directory
anymore, this is kind of annoying. By listing each archive file it will
use the parent directory of the archive as the root directory and only
include the archive, not any ancestor directories, in the artifact zip.