Remove code comment from artifact path
This was causing a problem. When this was setup as a glob path the
comment didn't cause any issues. But I think because it is just a simple
file path it doesn't pass it into the glob library. I guess the glob
library was able to strip the comment off? Or maybe not, either way it
prevents it from matching anything and is not necessary anymore.