Always make BUILD_DISTPATH a posix path
After `action-gh-release` was updated globs are handled differently.
This is because the glob library it uses was updated. The updated glob
library always interprets backslashes as escape characters. This means
if BUILD_DISTPATH has backslashes in it, `action-gh-release` wont match
it correctly and won't upload our archives. To fix this, we can make
sure to always use a posix path.