Add `--use-pep517` flag to `pip installs`
This forces the use of PEP 517 when using `pip install`. This requires
all dependencies to be installed via a wheel. With this flag, `pip` will
install the wheel dependency and build wheels for any package that
doesn't have one.

The motivation for this is that `cairocffi` doesn't have a wheel and
causes problems when installing via `pip`, but only when PyOxidizer
tries to install it. When installed with regular `pip` it works fine.
The `PyOxidizer` `pip` will error out on install. But when it installs
it via a wheel, there are no problems. So to get around this odd
behavior, I can add the `--use-pep517` flag into `PyOxidizer`'s starlark
script, and it will build the wheel first, then install, no problemo.

Since I will use the `--use-pep517` flag for `PyOxidizer`, I figure I
might as well enforce its use throughout the scriptopoly and install
scripts, wherever a `pip install` is used.