Avoid creating png chart with PyOxidizer build
It seems infeasible to get the png creation working with PyOxidizer.
There were even more __file__ issues in other dependencies that needed
fixing and there was also this issue:
https://github.com/indygreg/PyOxidizer/issues/434. That was on an older
version but I'm not interested in jumping through hoops just to find out
this is not possible with 'in-memory' loading. It works on the other
builds so thats good enough for me.