Remove cairosvg, don't render `png`s
When the binary was built on Github Actions, I couldn't successfully run
it locally. The problem was when it tried to render the png. This has
to do with the way `cairocffi` dynamically generates a file when it is
installed. This links to the cairo library on the system it is installed
in. It doesn't seem to work correctly when it is run on a different
system than it was installed on.

I don't have time to further troubleshoot this so I'm just going to drop
png rendering...for now.