Move to use PEP 621, keep poetry
This uses `pyproject.toml` to include project metadata, as described in
`PEP 621`. At some point this will be managed by `poetry`, for the
time being it is not. But it doesn't change very frequently (just the
version occasionally) so manually tweaking it as needed shouldn't be all
that big of a deal.

Doing this also allows me to remove most of `setup.py` but I still need
to keep the C extension building code. I also don't really think there
is a great way to replace the `setup.py build` or `setup.py build_ext`
with `pip` equivalents, so for now I'm going to keep them as is. As far
as I could tell, I would have to install the app into a temporary place
and maybe copy the files, or build wheels and extract the appropriate
files where I need them to be. Both of those approaches seem much more
complex than just calling `setup.py`. If I revisit this in the future
and there is a better option, then great, but right now there really
isn't.