Add 'Rich' text to monopoly's text output
This adds the Rich package to get nicer looking text output in the
terminal. Rich also comes with a spinner so the Spinner from utils is
no longer needed.

This also required adding a hash to PyGal's requirements-runtime.txt
entry because of how pip handles hashes. If one requirement has it, they
all must. In the process of doing this I ended up building a wheel of
the PyGal patch so that is what is now used to install PyGal. In the
end, this is probably an improvement.