Fix KB Interrupt issue with PyInstaller/Nuitka
For some reason, the KeyboardInterrupt error was being raised multiple
times, and because of this, the second one would raise an error that is
not handled and mess up the handling of the initial raise. This fixes
that by routing the keyboard interrupt signal to a Python function and
only raising the first interrupt that is encountered.