Use starmap_async instead to help on Windows
This switches to using starmap_async instead of starmap because on
Windows, starmap blocks the keyboard interrupt signal from getting
handled until it finishes. This is obviously not ideal as there can be
a huge delay from when the interrupt is triggered until it is handled.
By using starmap_async and also looping and sleeping while we wait for
it to complete, the interrupt is handled right away AND there is no
noticeable degradation in performance.