Improve status indicator on Windows
This improves the status indicator on Windows by using a custom class to
handle rendering. Using rich created a blinking effect across the whole
line  while the bouncing bars appeared to skip quite a lot. Not very
nice looking. This takes a different approach and brings back much of
the old Spinner code, while updating it to only handle Windows, and also
using the internals of rich to still create nice looking styles.